requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30

# Core utilities
python-dotenv>=1.0
pydantic>=2.0

# LangChain stack
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
httpx>=0.25
async-lru>=2.0.4

# Vector store and embeddings
faiss-cpu>=1.7
rank-bm25>=0.2.2

# NLP models
transformers>=4.40
sentence-transformers>=2.5

# Optional but commonly needed
numpy>=1.24
scikit-learn>=1.3
torch>=2.1; platform_system != "Darwin" or platform_machine != "arm64"
torch>=2.1; platform_system == "Darwin" and platform_machine == "arm64"
gradio>=4.31
